+++ docs/changelog.rst
@@ -29,3 +29,5 @@ Version 2.2
 * Added meta keywords and description to the main page
 * Added meta keywords using recipe titles and tages to the recipe detail page
 * Fixed bug of ratings not showing up on a recipe till a user voted on them even if another user had rated the recipe
+* Added a top recipe view that shows the top 20 recipes
+* Added a most recent recipes view that shows the 20 most recent recipes